backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.34.82
botocore==1.34.82
Brotli==1.0.9
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
cryptography==42.0.5
deprecation==2.1.0
distlib==0.3.6
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==3.0.10
gotrue==2.6.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.25.2
hyperframe==6.0.1
idna==3.7
importlib-metadata==7.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.3
packaging==24.1
pip==24.0
postgrest==0.16.9
psycopg2==2.9.9
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.0
realtime==1.0.6
requests==2.32.2
s3transfer==0.10.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
typing_extensions==4.12.2
urllib3==1.26.18
virtualenv==20.17.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
zipp==3.17.0
